requirements.txt
# Recommended: Python 3.10 or 3.11
numpy>=1.26,<3
nltk>=3.9,<4
tensorflow-cpu>=2.16,<2.20
jupyter>=1,<2
ipykernel>=6,<7
